+++ DOIT/src/main/java/it/unicam/cs/ids/DOIT/role/IPartecipationRequestHandler.java
@@ -5,9 +5,9 @@ import it.unicam.cs.ids.DOIT.partecipation_request.IPartecipationRequest;
 import java.util.Set;
 
 public interface IPartecipationRequestHandler {
-    void acceptPR(int idRole, int idProject) throws RoleException;
+    void acceptPR(Long idRole, Long idProject) throws RoleException;
 
-    void removePR(int idRole, int idProject, String description);
+    void removePR(Long idRole, Long idProject, String description);
 
-    Set<IPartecipationRequest> getPartecipationRequestsByTeam(int idProject);
+    Set<IPartecipationRequest> getPartecipationRequestsByTeam(Long idProject);
 }
